observability-system/metrics-proxy/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
pyyaml>=6.0